backend/requirements.txt
asgiref==3.8.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.13.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
